+++ packages/wiz-ui-next/src/components/base/inputs/number-input/index.ts
@@ -1,0 +1,1 @@
+export { default as WizNumberInput } from "./number-input.vue";