+++ packages/wiz-ui-next/src/components/base/popup-button-group/types.ts
@@ -1,7 +1,7 @@
 import { TIcon } from "@/components";
-export interface PopupButtonOption {
+export interface PopupButtonOption<T = number> {
   label: string;
-  value: number;
+  value?: T;
   exLabel?: string;
   icon?: TIcon;
   iconDefaultColor?: "green.800" | "gray.500";
@@ -9,13 +9,13 @@ export interface PopupButtonOption {
   onClick: () => void;
 }
 
-export type ButtonGroupItem =
-  | { kind: "button"; option: PopupButtonOption }
+export type ButtonGroupItem<T = number> =
+  | { kind: "button"; option: PopupButtonOption<T> }
   | { kind: "divider" }
   | {
       kind: "group";
       title: string;
-      items: ButtonGroupItem[];
+      items: ButtonGroupItem<T>[];
       groupDivider?: boolean;
       buttonDivider?: boolean;
     };